fastapi_app/requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
dictdiffer==0.9.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
Faker==17.6.0
fastapi==0.93.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
minio==7.1.13
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomlkit==0.11.6
tortoise-orm==0.19.3
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.20.0
